Fix/join-session (#71)
* WIP: Prevent users from joining other sessions

Currently so long as you have the URL to a session, you can join the
session and end it or send messages to it. This remains the case with
the existing code added in this commit, which only prevents overwriting
the values set on the session object.

To complete this work, we'd need to add logic to ensure that upon
attempting to join a session I don't belong to, I'm redirected to my
dashboard page.

* Added logic to socket handlers and routes to block non-participants.

Moved original code from join handler into a helper module for purposes of reuse.

* added logic to direct non-participants away from session UI

* Fixed error handling in Session.joinUser.

When we handle this in the model functions, we don't need to check again in sockets.js; we tackle
both the original issue and the race condition this way.

* equality operators: != --> !==

* ObjectId.equals() is better than !== for this purpose

* handle database errors correctly

* Use callback to handle error

* Removed dead code

Also some style fixes.