Refactoring for feature flag MUTED_SUBJECT_ALERTS
Since this feature flag relies on posthog data which is not currently
present in the backend postgresql database, this information could not
be used directly in the getNextVolunterToNotify query. Hence, this
commit refactors the query to return multiple volunteer candidates,
and the feature flag is then used in TwilioService.ts.