Revert back to getting one volunteer to notify
This reverts some of the changes made in commit:
77f2ead0bfb0b91ed12a15189fd7067a611f45a1

The motivation here is to reduce the scope of possible impact of
modifying a critical part of the backend: finding and notifiying
volunteers for new sessions. We keep the check for muted subject
alerts in TwilioService, but only use one candidate volunteer at
a time now from the reverted query getNextVolunteerToNotify.