Update moderation service to record instances of censorship
As part of the moderation work we're doing, we want to track which
  instances of censorship to get an idea of how often it happens and
  what types of messages are being censored.

We hope to analyze chat transcripts with various moderation techniques
(mainly AI moderators) and compare whether this instance of censorship
was justified or not (e.g. "Kill the pig. Cut her throat. Spill her blood."
is maybe an okay thing to type when talking about a Lord of the Flies essay).